+++ packages/riverpod/lib/src/core/override.dart
@@ -8,9 +8,14 @@ part of '../framework.dart';
 @publicInMisc
 sealed class Override {}
 
-sealed class _ProviderOverride implements Override {}
+/// An object used by [ProviderContainer]/`ProviderScope` to override the behavior
+/// of a provider for part of the application.
+@visibleForTesting
+sealed class ProviderOverride implements Override {}
 
-extension on _ProviderOverride {
+/// Extension methods for [ProviderOverride].
+@visibleForTesting
+extension ProviderOverrideX on ProviderOverride {
   /// The provider that is overridden.
   $ProviderBaseImpl<Object?> get origin {
     final that = this;
@@ -19,7 +24,9 @@ extension on _ProviderOverride {
       $ProviderOverride() => that.origin,
     };
   }
+}
 
+extension on ProviderOverride {
   /// The new provider behavior.
   $ProviderBaseImpl<Object?> get providerOverride {
     final that = this;
@@ -30,9 +37,14 @@ extension on _ProviderOverride {
   }
 }
 
-sealed class _FamilyOverride implements Override {}
+/// An object used by [ProviderContainer]/`ProviderScope` to override the behavior
+/// of a family for part of the application.
+@visibleForTesting
+sealed class FamilyOverride implements Override {}
 
-extension on _FamilyOverride {
+/// Extension methods for [FamilyOverride].
+@visibleForTesting
+extension FamilyOverrideX on FamilyOverride {
   /// The provider that is overridden.
   Family get from {
     final that = this;
@@ -40,6 +52,26 @@ extension on _FamilyOverride {
       Family() => that,
       $FamilyOverride() => that.from,
     };
+  }
+}
+
+/// Extension methods for `List<Override>`.
+@visibleForTesting
+extension ListOverrideX on List<Override> {
+  /// Returns `true` if the list already contains an [Override] for the family or provider that [override] overrides, `false` otherwise.
+  bool containsOverride(Override override) {
+    return any((o) {
+      return switch ((o, override)) {
+        (
+          ProviderOverride(:final origin),
+          ProviderOverride(origin: final otherOrigin),
+        ) =>
+          origin == otherOrigin,
+        (FamilyOverride(:final from), FamilyOverride(from: final otherFrom)) =>
+          from == otherFrom,
+        _ => false,
+      };
+    });
   }
 }
 
@@ -53,7 +85,7 @@ extension on _FamilyOverride {
 /// - [ProviderContainer], which uses this object.
 /// - `overrideWithValue`, which creates a [$ProviderOverride].
 @publicInCodegen
-class $ProviderOverride implements _ProviderOverride {
+class $ProviderOverride implements ProviderOverride {
   /// Override a provider
   $ProviderOverride({required this.origin, required this.providerOverride});
 
@@ -94,7 +126,7 @@ class TransitiveProviderOverride implements $ProviderOverride {
 /// to override the behavior of a "family" for part of the application.
 @internal
 @publicInCodegen
-abstract class $FamilyOverride implements _FamilyOverride {
+abstract class $FamilyOverride implements FamilyOverride {
   /// Do not use: Internal object to used by [ProviderContainer]/`ProviderScope`
   /// to override the behavior of a "family" for part of the application.
   factory $FamilyOverride({